バッチのスライス処理を外部モジュール化
